move config-related code to config.py
