+++ cognite_toolkit/cognite_modules/examples/cdf_data_pipeline_files_valhall/functions/fu_context_files_oid_fileshare_annotation/handler.py
@@ -9,7 +9,6 @@ from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 from typing import Any, Optional
 
-import yaml
 from cognite.client import ClientConfig, CogniteClient
 from cognite.client.credentials import OAuthClientCredentials
 from cognite.client.data_classes import (
@@ -21,8 +20,9 @@ from cognite.client.data_classes import (
     FileMetadataUpdate,
 )
 from cognite.client.data_classes.contextualization import DiagramDetectResults
-from cognite.client.exceptions import CogniteAPIError
 from cognite.client.utils._text import shorten
+
+from .config import AnnotationConfig, load_config_parameters
 
 # P&ID original file defaults
 ORG_MIME_TYPE = "application/pdf"
@@ -41,33 +41,6 @@ CREATING_APPVERSION = "1.0.0"
 # Other constants
 ASSET_MAX_LEN_META = 10000
 ISO_8601 = "%Y-%m-%d %H:%M:%S"
-
-
-@dataclass
-class AnnotationConfig:
-    extpipe_xid: str
-    debug: bool
-    run_all: bool
-    doc_limit: int
-    doc_data_set_ext_id: str
-    doc_type_meta_col: str
-    p_and_id_doc_type: str
-    asset_root_ext_ids: list[str]
-    match_threshold: float
-
-    @classmethod
-    def load(cls, data: dict[str, Any]) -> AnnotationConfig:
-        return cls(
-            extpipe_xid=data["ExtractionPipelineExtId"],
-            debug=data["debug"],
-            run_all=data["runAll"],
-            doc_limit=data["docLimit"],
-            doc_data_set_ext_id=data["docDataSetExtId"],
-            doc_type_meta_col=data["docTypeMetaCol"],
-            p_and_id_doc_type=data["pAndIdDocType"],
-            asset_root_ext_ids=data["assetRootExtIds"],
-            match_threshold=data["matchThreshold"],
-        )
 
 
 @dataclass
@@ -90,27 +63,11 @@ class Entity:
 
 def handle(data: dict, client: CogniteClient) -> dict:
     config = load_config_parameters(client, data)
-    annotate_p_and_id(client, config)
+    annotate_pnid(client, config)
     return {"status": "succeeded", "data": data}
 
 
-def load_config_parameters(cognite_client: CogniteClient, function_data: dict[str, Any]) -> AnnotationConfig:
-    """Retrieves the configuration parameters from the function data and loads the configuration from CDF."""
-    if "ExtractionPipelineExtId" not in function_data:
-        raise ValueError("Missing key 'ExtractionPipelineExtId' in input data to the function")
-
-    extpipe_xid = function_data["ExtractionPipelineExtId"]
-    try:
-        extpipe_config = cognite_client.extraction_pipelines.config.retrieve(extpipe_xid)
-    except CogniteAPIError:
-        raise RuntimeError(f"Not able to retrieve pipeline config for extraction pipeline: {extpipe_xid!r}")
-
-    data = yaml.safe_load(extpipe_config.config)["data"]
-    data["ExtractionPipelineExtId"] = extpipe_xid
-    return AnnotationConfig.load(data)
-
-
-def annotate_p_and_id(cognite_client: CogniteClient, config: AnnotationConfig) -> None:
+def annotate_pnid(cognite_client: CogniteClient, config: AnnotationConfig) -> None:
     """
     Read configuration and start P&ID annotation process by
     1. Reading files to annotate
@@ -125,11 +82,7 @@ def annotate_p_and_id(cognite_client: CogniteClient, config: AnnotationConfig) -
     Args:
         cognite_client: An instantiated CogniteClient
         config: A dataclass containing the configuration for the annotation process
-
     """
-
-    print("Initiating Annotation[INFO]  of P&ID")
-
     for asset_root_ext_id in config.asset_root_ext_ids:
         try:
             all_files, filer_to_process = get_files(
@@ -155,7 +108,6 @@ def annotate_p_and_id(cognite_client: CogniteClient, config: AnnotationConfig) -
                     annotation_list,
                     config,
                 )
-
             msg = (
                 f"Annotated P&ID files for asset: {asset_root_ext_id} number of files annotated: {annotated_count}, "
                 f"file not annotaded due to errors: {error_count}"
